requirements.txt
bson==0.4.1
Flask==0.10.1
Flask-Assets==0.11
Flask-Cors==2.1.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mongo==0.2.0
pymongo==3.1
pytz==2015.7
requests==2.8.1
six==1.10.0
webassets==0.11.1
Werkzeug==0.11.2
wheel==0.24.0
